Adds auction total display and cloud function
The running total text was added to the AuctionTeam component as a single line of test.  The new cloud function uses the onUpdate trigger to fire each time the 'owner' property of in the teams node of a league is changed.  The cloud function cycles through each team in the league and calculates each user's total bid as well as the grand total and calculates the use tax if applicable.